requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
django-filter==2.2.0
djangorestframework==3.10.2
jsonpickle==1.2
Markdown==3.1.1
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.2
six==1.14.0
sqlparse==0.3.0
